+++ addon/cards/markdown-card_html.js
@@ -1,3 +1,4 @@
+/* jshint ignore:start */
 // this has to be a node style module as it's imported by the front end and ember.
 // a card has to have an editor object, and an HTML object. In the future we might have AMP and Text (for FTS) renderers.
 
@@ -12,3 +13,4 @@ module.exports = {
     }
 };
 
+/* jshint ignore:end */
