+++ app/styles/ghost-toolbar-blockitem.css
@@ -2,23 +2,6 @@
 @import "globals.css";
 
 
-:root {
-    /* Colours */
-    --darkgrey: #242628;
-    --midgrey: #7d878a;
-    --lightgrey: #e2edf2;
-    --blue: #5ba4e5;
-    --red: #e25440;
-    --orange: #f2a925;
-    --green: #9fbb58;
-    /* Style values */
-    --border-radius: 4px;
-    --font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
-    --font-family-mono: Consolas, "Liberation Mono", Menlo, Courier, monospace;
-
-    --button-size: 48px;
-}
-
 
 
 .toolbar-block {
