+++ bower.json
@@ -1,8 +1,11 @@
 {
   "name": "ghost-editor",
   "dependencies": {
-    "ember": "~2.8.0",
+    "ember": "components/ember#release",
     "ember-cli-shims": "0.1.1",
     "codemirror": "~5.15.0"
+  },
+  "resolutions": {
+    "ember": "release"
   }
 }
