--- bower.json.ember-try
@@ -1,7 +1,0 @@
-{
-  "name": "ghost-editor",
-  "dependencies": {
-    "ember": "~2.8.0",
-    "ember-cli-shims": "0.1.1"
-  }
-}
