+++ tests/integration/components/ghost-editor-test.js
@@ -1,0 +1,24 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('ghost-editor', 'Integration | Component | ghost editor', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{ghost-editor}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#ghost-editor}}
+      template block text
+    {{/ghost-editor}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
