--- tests/unit/helpers/if-equals-test.js
@@ -1,12 +1,0 @@
-
-import { ifEquals } from 'dummy/helpers/if-equals';
-import { module, test } from 'qunit';
-
-module('Unit | Helper | if equals');
-
-// Replace this with your real tests.
-test('it works', function(assert) {
-  let result = ifEquals([42]);
-  assert.ok(result);
-});
-
