+++ xtestsx/dummy/app/controllers/application.js
@@ -1,0 +1,5 @@
+import Ember from 'ember';
+
+export default Ember.Controller.extend({
+
+});
