+++ xtestsx/todo.txt
@@ -1,0 +1,2 @@
+slashmenu:
+    when close restore keys
